+++ platform/ios/src/MGLFaux3DUserLocationAnnotationView.h
@@ -1,0 +1,7 @@
+#import <UIKit/UIKit.h>
+#import "MGLUserLocationAnnotationView.h"
+
+@interface MGLFaux3DUserLocationAnnotationView : MGLUserLocationAnnotationView
+
+@end
+